Made changes to the profile page and account component.

+++ api/models/card.go
@@ -13,7 +13,7 @@ type Card struct {
 	Brand        string             `bson:"brand" validate:"required"`
 	SetName      string             `bson:"set_name" validate:"required"`
 	Rarity       string             `bson:"rarity" validate:"required"`
-	SellingPrice uint               `bson:"selling_price" validate:"required"`
+	SellingPrice float64            `bson:"selling_price" validate:"required"`
 	UploadedAt   time.Time          `bson:"uploaded_at"`
 	CardID       string             `bson:"card_id"`
 	Image        string             `bson:"image_url"`
